Initial timestamper
The timestamper encodes a timestamp as a series of nodes ready to be used
to learn about a person's habits. It uses a total of 47 nodes per
timestamp, and has 24 per day, 7 a week, 4 a month, and 12 a year.

Hopefully this will allow pretty fast learning, taking into account
people's schedules.